requirements.txt
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.0.0
transformers==4.30.2
numpy==1.21.4
pandas==1.5.3
pytokenizations==0.8.4
pytorch-lightning==2.0.4
skweak==0.3.3
spacy==3.5.4
regex==2023.6.3
scikit-learn==1.2.2
beautifulsoup4==4.12.2

